refactor: split ssh cmd into exec and scp (#996)
